Aligned test package structure with the other modules
