+++ types-jackson/src/test/java/dk/cloudcreate/essentials/jackson/model/ProductId.java
@@ -14,25 +14,27 @@
  * limitations under the License.
  */
 
-package dk.cloudcreate.essentials.types;
+package dk.cloudcreate.essentials.jackson.model;
+
+import dk.cloudcreate.essentials.types.*;
 
 import java.util.UUID;
 
-public class CustomerId extends CharSequenceType<CustomerId> implements Identifier {
-    public CustomerId(CharSequence value) {
+public class ProductId extends CharSequenceType<ProductId> implements Identifier {
+    public ProductId(CharSequence value) {
         super(value);
     }
 
-    public static CustomerId of(CharSequence value) {
-        return new CustomerId(value);
+    public static ProductId of(CharSequence value) {
+        return new ProductId(value);
     }
 
-    public static CustomerId ofNullable(CharSequence value) {
-        return value != null ? new CustomerId(value) : null;
+    public static ProductId ofNullable(CharSequence value) {
+        return value != null ? new ProductId(value) : null;
     }
 
-    public static CustomerId random() {
+    public static ProductId random() {
         // You can use any random Id generator like e.g. https://github.com/codahale/time-id
-        return new CustomerId(UUID.randomUUID().toString());
+        return new ProductId(UUID.randomUUID().toString());
     }
 }
